Move any Meta modifying functions, into Meta_Helper
This contains no actual logic change, simply cosmetically moving code from one place to another, and making sure all related code is updated
